Game Update
A new version of the Asteroids game has been released. It is still in its early access so there is no start menu.
Impressively this is just over half the size of the origional code while also keeping functionality.

Controls:
WASD or Arrow keys for movement
F to reset the game when you loose

Try to avoid the asteroids!