Vsync Update!
With the latest version of the game_engine_50122, it now supports vsync with a smooth 60FPS gameplay.
Additionally, the complexity of the code has been changed to support the newer method of f-strings. However, this does mean that the required version of python is 3.6.
Furthermore, a higher level of documentation has been added to the source code for all users to be able to easily understand all the functions.